PolynomialMatrixEquations.jl

This package contains algorithms to solve Quadrilateral Matrix Equations in the form of

\[A_0 + A_1 X + A_2 X^2 = \bm{0}.\]

Cyclic Reduction

This algorithm adapted from Bini et al.[1] solves the second order matrix equation iteratively for $X$. It is implemented both for $A_0$, $A_1$ and $A_2$ dense or sparse (i.e. SparseMatrixCSC). In the sparse version, a special packing is performed that maximally exploits the column and row sparsity of the $A$ matrices.

Generalized Schur Decomposition Solver

The polynomial matrix equation

\[A_0 + A_1 X + A_2 X^2 = \bm{0}\]

can be rewritten in the form

\[\begin{bmatrix} 0 & A_2 \\ I & 0 \end{bmatrix} \begin{bmatrix} I \\ X \end{bmatrix} X = -\begin{bmatrix} A_0 & A_1 \\ 0 & I \end{bmatrix} \begin{bmatrix} I \\ X \end{bmatrix}\]

In practive, it is possible to eliminate empty columns in $A_0$, $A_1$ and $A_2$ (See [2-4] for further info). More generally, this solver finds a solution to the following equation:

\[D \begin{bmatrix} I \\ X_2 \end{bmatrix} X_1 = E \begin{bmatrix} I \\ X_2 \end{bmatrix}\]

where matrices $X_1$ and $X_2$ contain columns from the solution matrix $X$.

Interface

FastLapackInterface.Workspace — Type

Workspace(a0, a1, a2)
Workspace(d, e, n)

Will return the right workspace for solve! on a given set of a0, a1 and a2, or matrices d, e and number of stable solutions n. If DenseMatrices are passed in, or d and e, it will return a GSSolverWs, else a CRSolverWs.

PolynomialMatrixEquations.solve — Function

solve

Will first construct the right Workspace and then call solve!.

PolynomialMatrixEquations.solve! — Function

solve!(ws, a0, a1, a2; kwargs...)

In place solve of the Quadrilateral Matrix Equation $a0 + a1 * x + a2 * x * x = 0$ for x and returns x.

Uses:

Generalized Schur when ws is a GSSolverWs and a0, a1 and a2 are DenseMatrices

Cyclic Reduction when ws is a CRSolverWs and at least a0 and a2 are SparseMatrixCSC

Further kwargs:

tolerance <1e-8>: convergence tolerance when using cyclic reduction, or tolerance for stable eigenvalues of the schur decomposition (eigval > 1 + tolerance is stable)

max_iterations <100>: maximum iterations allowed in cyclic reduction

Examples:

julia> using PolynomialMatrixEquations

julia> const PME = PolynomialMatrixEquations
PolynomialMatrixEquations

julia> a0 = [1.21 0.0 0.0
             0.27 0.0 0.0
             1.93 0.0 0.0];

julia> a1 = [0.81  -0.68 0.22
             -1.12 0.14 -0.88
             -1.1 0.12 -1.14];

julia> a2 = [0.0 -0.42 -0.42
             0.0 0.29 -1.36
             0.0 0.23 -1.09];

julia> ws2 = PME.Workspace(a0, a1, a2)
PolynomialMatrixEquations.GSSolverWs{Float64}
  tmp1: 1×1 Matrix{Float64}
  tmp2: 1×1 Matrix{Float64}
  g1: 1×1 Matrix{Float64}
  g2: 2×1 Matrix{Float64}
  x: 3×3 Matrix{Float64}
  d: 3×3 Matrix{Float64}
  e: 3×3 Matrix{Float64}
  luws1: FastLapackInterface.LUWs
  luws2: FastLapackInterface.LUWs
  schurws: FastLapackInterface.GeneralizedSchurWs{Float64}

julia> PME.solve!(ws2, a0, a1, a2)
3×3 Matrix{Float64}:
 -0.511005  0.0  0.0
  5.72536   0.0  0.0
  4.29884   0.0  0.0

julia> using SparseArrays

julia> a0 = sparse([1.21 0.0 0.0
                    0.27 0.0 0.0
                    1.93 0.0 0.0]);

julia> a1 = sparse([0.81  -0.68 0.22
                    -1.12 0.14 -0.88
                    -1.1 0.12 -1.14]);

julia> a2 = sparse([0.0 -0.42 -0.42
                    0.0 0.29 -1.36
                    0.0 0.23 -1.09]);

julia> ws2 = PME.Workspace(a0, a1, a2)
PolynomialMatrixEquations.CRSolverWs{Float64, FastLapackInterface.LUWs, SparseMatrixCSC{Float64, Int64}}
  linsolve_ws: FastLapackInterface.LUWs
  ahat1: 3×3 Matrix{Float64}
  a1copy: 3×3 Matrix{Float64}
  x: 3×3 SparseMatrixCSC{Float64, Int64} with 3 stored entries
  m: 6×6 Matrix{Float64}
  m1: 3×6 Matrix{Float64}
  m2: 6×3 Matrix{Float64}

julia> PME.solve!(ws2, a0, a1, a2)
3×3 SparseMatrixCSC{Float64, Int64} with 3 stored entries:
 -0.511005   ⋅    ⋅
  5.72536    ⋅    ⋅
  4.29884    ⋅    ⋅

Workspaces

PolynomialMatrixEquations.GSSolverWs — Type

GSSolverWs

Workspace for solving with the Generalized Schur solver. GSSolverWs(A) with A an example Matrix.

PolynomialMatrixEquations.CRSolverWs — Type

CRSolverWs

Workspace used for solving with the cyclic reduction algorithm of Bini et al..

References

D.A. Bini, B. Iannazzo, B. Meini, Numerical Solution of Algebraic Riccati Equations, SIAM Book Series Fundamentals of Algorithms, 2012.

P. Klein Using the generalized Schur form to solve a linear rational expectations model, Journal of Economic Dynamics and Control, 2000.

C. Sims Solving rational expectations models, Computational Economics, 2002.

N. J. Higham and H. Kim Numerical analysis of a quadratic matrix equation, Journal of Numerical Analysis, 2000.