PolynomialMatrixEquations.jl

This package contains algorithm to solve various polynomial matrix equations.

Cyclic Reduction

This algorithm adapted from Bini et al. [1] solves the second order matrix equation iteratively:

\[A_0 + A_1 X + A_2 X^2 = \bm{0}\]

for $X$. It is implemented both for $A_0$, $A_1$ and $A_2$ dense or sparse (i.e. SparseMatrixCSC). In the sparse version, a special packing is performed that maximally exploits the column and row sparsity of the $A$ matrices.

PolynomialMatrixEquations.CRSolverWs — Type

CRSolverWs

Workspace used for solving with the cyclic reduction algorithm. Can be constructed as CRSolverWs(n) with n the leading dimension of the matrics $A_0$, $A_1$ and $A_2$, i.e. the number of equations.

Missing docstring.

Missing docstring for PolynomialMatrixEquations.solve!(::CRSolverWs{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}) where {T<:AbstractFloat}. Check Documenter's build log for details.

Generalized Schur Decomposition Solver

The polynomial matrix equation

\[A_0 + A_1 X + A_2 X^2 = \bm{0}\]

can be rewritten in the form

\[\begin{bmatrix} 0 & A_2 \\ I & 0 \end{bmatrix} \begin{bmatrix} I \\ X \end{bmatrix} X = -\begin{bmatrix} A_0 & A_1 \\ 0 & I \end{bmatrix} \begin{bmatrix} I \\ X \end{bmatrix}\]

In practive, it is possible to eliminate empty columns in $A_0$, $A_1$ and $A_2$ (See Klein, 2000, and . More generally, this solver finds a solution to the following equation:

\[D \begin{bmatrix} I \\ X_2 \end{bmatrix} X_1 = E \begin{bmatrix} I \\ X_2 \end{bmatrix}\]

where matrices $X_1$ and $X_2$ contain columns from the solution matrix $X$.

PolynomialMatrixEquations.GSSolverWs — Type

GSSolverWs

Workspace for solving with the GeneralizedSchur solver. Can be constructed using a matrix and the number of stable solutions, i.e. GSSolverWs(A, n) with A an example Matrix and n the number of solutions.

PolynomialMatrixEquations.solve! — Method

solve!([ws::GSSolverWs,] d::Matrix, e::Matrix, n1::Int64, qz_criterium)

The solution is returned in ws.g1 and ws.g2. d and e are mutated during the solving. n1 determines how many stable solutions should be found.

References

D.A. Bini, B. Iannazzo, B. Meini, *Numerical Solution of Algebraic

Riccati Equations*, SIAM Book Series Fundamentals of Algorithms, 2012.

P. Klein Using the generalized Schur form to solve a linear rational expectations model, Journal of Economci Dynamics and Control, 2000.

C. Sims Solving rational expectations models, Computational Economics, 2002.