PolynomialMatrixEquations.jl

This package contains algorithm to solve various polynomial matrix equations.

Cyclic Reduction

This algorithm adapted from Bini et al. [1] solves the second order matrix equation iteratively:

\[A_0 + A_1 X + A_2 X^2 = \bm{0}\]

for $X$. It is implemented both for $A_0$, $A_1$ and $A_2$ dense or sparse (i.e. SparseMatrixCSC). In the sparse version, a special packing is performed that maximally exploits the column and row sparsity of the $A$ matrices.

PolynomialMatrixEquations.CRSolverWs — Type

CRSolverWs

Workspace used for solving with the cyclic reduction algorithm. Can be constructed as CRSolverWs(n) with n the leading dimension of the matrics $A_0$, $A_1$ and $A_2$, i.e. the number of equations.

PolynomialMatrixEquations.solve! — Method

solve!([ws::CRSolverWs, ],
       x::AbstractMatrix,
       a0::AbstractMatrix,
       a1::AbstractMatrix,
       a2::AbstractMatrix;
       tolerance=1e-8,
       iterations=100)

Solves the quadratic matrix equation a0 + a1*x + a2*x*x = 0, using the cyclic reduction method from Bini et al. (???). If a0 and a2 are SparseMatrixCSC, a variation will be used that optimally packs the equations. a1 will always be used (i.e. potentially converted) as standard Matrix.

The solution is returned in x. In case of nonconvergency, x is set to NaN and UndeterminateSystemExcpetion or UnstableSystemException is thrown.

During the solving, x, a1 and ws are mutated. Use solve(a0, a1, a2) for a non-mutating version.

Example

julia> using PolynomialMatrixEquations

julia> using LinearAlgebra

julia> n = 3;

julia> ws = CRSolverWs(n);

julia> a0 = [0.5 0 0; 0 0.5 0; 0 0 0];

julia> a1 = Matrix(1.0I, n, n);

julia> a2 = [0 0 0; 0 0 0; 0 0 0.8];

julia> x = zeros(n,n);

julia> PolynomialMatrixEquations.solve!(ws, x, a0, a1, a2, tolerance = 1e-8, iterations = 50)
3×3 Matrix{Float64}:
 -0.5  -0.0  -0.0
 -0.0  -0.5  -0.0
 -0.0  -0.0  -0.0

Generalized Schur Decomposition Solver

The polynomial matrix equation

\[A_0 + A_1 X + A_2 X^2 = \bm{0}\]

can be rewritten in the form

\[\begin{bmatrix} 0 & A_2 \\ I & 0 \end{bmatrix} \begin{bmatrix} I \\ X \end{bmatrix} X = -\begin{bmatrix} A_0 & A_1 \\ 0 & I \end{bmatrix} \begin{bmatrix} I \\ X \end{bmatrix}\]

In practive, it is possible to eliminate empty columns in $A_0$, $A_1$ and $A_2$ (See Klein, 2000, and . More generally, this solver finds a solution to the following equation:

\[D \begin{bmatrix} I \\ X_2 \end{bmatrix} X_1 = E \begin{bmatrix} I \\ X_2 \end{bmatrix}\]

where matrices $X_1$ and $X_2$ contain columns from the solution matrix $X$.

PolynomialMatrixEquations.GSSolverWs — Type

GSSolverWs

Workspace for solving with the GeneralizedSchur solver. Can be constructed using a matrix and the number of stable solutions, i.e. GSSolverWs(A, n) with A an example Matrix and n the number of solutions.

PolynomialMatrixEquations.solve! — Method

solve!([ws::GSSolverWs,] d::Matrix, e::Matrix, n1::Int64, qz_criterium)

The solution is returned in ws.g1 and ws.g2. d and e are mutated during the solving. n1 determines how many stable solutions should be found.

References

D.A. Bini, B. Iannazzo, B. Meini, *Numerical Solution of Algebraic

Riccati Equations*, SIAM Book Series Fundamentals of Algorithms, 2012.

P. Klein Using the generalized Schur form to solve a linear rational expectations model, Journal of Economci Dynamics and Control, 2000.

C. Sims Solving rational expectations models, Computational Economics, 2002.